IB/mlx4: Align query device response
This patch explicitly aligns 'mlx4_uverbs_ex_query_device_resp' layout
to u64 and use the correct response's length when copying the data to
the user area.

The uAPI should be u64 aligned, currently on 64 bit systems this is done
implicitly as of the next u64 field in the rss_caps structure.

Legacy user space 32 bit system application in any case is broken with
64 bit system kernel before this change.

A matching change will be done in the user area for future use also on
32 bit systems.

In addition, we move the structure to the uAPI file where it's expected
to be from day one.

Fixes: 09d208b258a2 ("IB/mlx4: Add report for RSS capabilities by vendor channel")
Reviewed-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>