RDMA: Use rdma_query_gid() instead of ib_get_cached_gid()
rxe driver is interested in querying the GID attributes, change it to
use rdma_get_gid_attr() which performs similar task as that of
ib_get_cached_gid().

Rest of the callers of ib_get_cached_gid() are not interested in
GID attributes. Therefore shift the callers to use rdma_query_gid().

This further makes ib_get_cached_gid() redundant API. Therefore it is
removed.

Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>