IB/ipoib: Fix wqe initialized param on ipoib set mode to connected
In IPoIB connected mode the WQE opcode can be IB_WR_SEND only (stateless
offloads are supported only in datagram mode), but it wasn't initialized.
It leads to unpredictable CQE errors which were triggered by
ipoib_cm_send() flow.

In order to fix it, we initialize the wqe opcode in the set mode function
to be IB_WR_SEND.

Reviewed-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Shalom Lagziel <shaloml@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>