net: Make RX-FCS and LRO mutually exclusive
LRO and RX-FCS offloads cannot be enabled at the same time since it is
not clear what should happen to the FCS of each coalesced packet.
The FCS is not really part of the TCP payload, hence cannot be merged
into one big packet. On the other hand, providing one big LRO packet
with one FCS contradicts the RX-FCS feature goal.

Use the fix features mechanism in order to prevent intersection of the
features and drop LRO in case RX-FCS is requested.

Enabling RX-FCS while LRO is enabled will result in:
$ ethtool -K ens6 rx-fcs
Actual changes:
large-receive-offload: off [requested on]
rx-fcs: on

Issue: 1309359
Change-Id: Ie8660fdd938ae696a3a0f8fb9f53672dd9be3284
Signed-off-by: Gal Pressman <galp@mellanox.com>