IB/uverbs: Introduce MPLS steering match filter
Adding a new MPLS steering match filter that could match against
a single MPLS tag field.

Since the MPLS header can reside in different locations in the packet's
protocols stack as well as be encapsulated with a tunnel protocol, it
is required to know the exact location of the header in the protocol
stack.
Therefore, when including the MPLS protocol spec in the specs list,
it is mandatory to provide the list in an ordered manner, such
that represents the actual headers order in a matching packet.

Vendor drivers who process the spec list and apply the matching rule
should treat the position of the MPLS spec in the spec list as the
actual location of the MPLS label in the packet's protocol stack.

Reviewed-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>