RDMA/core: Remove obsolete UCM module
In commit 357d23c811a7 ("Remove the obsolete libibcm library")
in rdma-core [1], we removed obsolete library which used the
/dev/infiniband/ucmX interface. As it can be seen from
the discussion in that PR, all major distributions dropped
that library long time ago.

It is better to delete this interface from the kernel too,
before syzkaller team invests time and energy to harden
this unused interface.

[1] https://github.com/linux-rdma/rdma-core/pull/279

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>