IB/mlx5: Avoid adding mkeys to the cache when limit is zero
Limit zero for a cache order means this order should be empty and not used.
This is done for VFs and representors.

When the limit is zero, we expect mkey allocation from the cache to fail since
it's empty. Due to a bug, the cache was populated after hitting a cache miss in
the first allocation, and thus future allocations found mkeys in the cache.

Fix this by avoiding cache population when the limit is zero.

Fixes: e126ba97dba9 ("mlx5: Add driver for Mellanox Connect-IB adapters")
Reviewed-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Majd Dibbiny <majd@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>