net/mlx5: E-Switch, Fix adding rules without a match on outer headers
Before this commit the driver always tells the FW there is a match on
outer headers but it could be a user didn't provide any match.
This commit checks if there is anything to match and only then set the bit
that tells the FW that there is a matching on outer headers.

Issue: 1338214
Change-Id: Iba5ab9bc382a3d8784c6fa34c9c8f0824ca08553
Fixes: 3d80d1a2f59a ("net/mlx5: E-Switch, Add API to configure rules for the offloaded mode")
Signed-off-by: Roi Dayan <roid@mellanox.com>