net/mlx5e: Use hint to resolve route in multipath mode
As part of creating a vxlan header we resolve the route and neighbour
in order to get the destination mac.

In multipath mode we could get the peer port as it is valid according
to the routing table but is the wrong mac address for the current rule
we want to offload so we use a hint to get the correct mac address.

Issue: 1360565
Change-Id: I6a03f6820f0d1aa3c5d33f450df6b80e4ea5edad
Signed-off-by: Shahar Klein <shahark@mellanox.com>