RDMA/cma: Zero out qp and ah attribute
Commit given in fixes tag introduced an accurate check to validate
device, port, index by referring to the cache layer for querying GIDs
for all link layers (IB, RoCE and iWarp).

When rdmacm tries to modify the QP to RTR state for kernel consumers,
qp and ah attributes are uninitialized. Each transport layer (IB/iWarp)
initializes them depending on transport type.
However qp ah_attr are not used for iWarp and remained uninitialized,
which is further used in ib_query_gid() call. This results into a
failure to query the GID due to an invalid GID index coming from
the uninitialized stack memory.
This is reported and discussed in thread [1].

This fix zero out the qp and ah attributes to avoid unitialized GID
index.

[1] https://www.spinics.net/lists/linux-rdma/msg63951.html

Fixes: f35faa4ba956 ("IB/core: Simplify ib_query_gid to always refer to cache")
Tested-by: Raju Rangoju <rajur@chelsio.com>
Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>