net/mlx5: Break encap/decap into two separated flow table creation flags
Today we are able to attach encap and decap actions only to the FDB.
In preparation to enable those actions on the NIC flow tables, break
the single flag into two. Those flags control whatever a decap
or encap operations can be attached to the flow table created, this
information is decided at the flow table creation and not when the
actual flow rule is created. Current FW support only decap on the NIC RX
flow tables and encap on NIC TX.

Signed-off-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>