net: sched: cls: Fix offloading when ingress dev is vxlan
When using a vxlan device as the ingress dev, we count it as a
"no offload dev", so when such a rule comes and err stop is true,
we fail early and don't try the egdev route which can offload it
through the egress device.

Fix that by not calling the block offload if one of the devices
attached to it is not offload capable, but make sure egress on such case
is capable instead.

Issue: 1360599
Change-Id: I5be4a02bacee22d043dde435bfe4a4c85bd6b53a
Fixes: caa7260156eb ("net: sched: keep track of offloaded filters [..]")
Signed-off-by: Paul Blakey <paulb@mellanox.com>