IB/uverbs: Don't use the address vector's port number during modify_qp
During uverbs' modify_qp, the port type is inferred using the port
number in ib_uverbs_qp_dest struct (address vector). This port number is
used to access that port's type info in the port_immutable array.

According to IB spec (version 1.3) port number is not mentioned as a
part of the AV which is used in modify QP. Using the port number field
from the AV therefore leads to accesses to non-allocated memory when
inferring the port type.

Instead, the port number from ib_qp->real_qp should be used to infer the
port type.

The commit cited below overlooked this portion of the fix.

Cc: <stable@vger.kernel.org> # 4.13
Fixes: 498ca3c82a7b ("IB/core: Avoid accessing non-allocated memory when inferring port type")
Signed-off-by: Guy Levi <guyle@mellanox.com>
Reviewed-by: Jack Morgenstein <jackm@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>