IB/core: Free GID table entry during GID deletion
If the GID table entry has no active users, which means its
kref count is 1. In such case, there is no need to schedule
GID deletion work. It can be completed in the caller's blocking
context in del_gid() directly.

This helps to eliminate issues when GID change is requested as part of
MAC address change or bonding event change where expectation is to
replace the GID almost immediately.

Fixes: b150c3862d21 ("IB/core: Introduce GID entry reference counts")
Reviewed-by: Daniel Jurgens <danielj@mellanox.com>
Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>