net/mlx5: Disable reuse flow rule
The upper layer can provide a rule where the output vport destination
is repeated multiple times, e.g under VF port mirroring scheme.

The current FS core implementation reuses the same flow rule, hence the
total flow rule number and the fte children number will not be the same.
That will cause some issues. And the benefit to reuse the flow rule is
not big. So we remove the related code in this patch.

Issue: 1416097

Change-Id: I00c01521d3519c606853bd076e65f829af85b503
Fixes: 74491de93712 ("net/mlx5: Add multi dest support")
Signed-off-by: Chris Mi <chrism@mellanox.com>