RDMA/mlx5: Don't leak UARs in case of free fails
The failure in releasing one UAR doesn't mean that we can't continue to
release rest of system pages, so don't return too early.

As part of cleanup, there is no need to print warning if mlx5_cmd_free_uar()
fails because such warning will be printed as part of mlx5_cmd_exec().

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>