net/mlx5e: Fix vport admin state to follow port state
In addition to port state set, driver shall set the vport state. This is
essential since a driver request for port state, does not guarantee an
actual change in the physical state of the port. However the vport is
fully controlled by the driver and must be changed on the specified
scenarios.

Add two helper functions in order to control the port and vport up/down
state.

Issue: 1365387
Change-Id: I04a853b4b57716e41515d175dcefd580d79426fd
Fixes: 63bfd399de55 ("net/mlx5e: Send PAOS command on interface up/down")
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>